refactor: tag 관련 로직 훅으로 분리 및 재사용
